rust: move apps/bitcoin to hww/api/bitcoin
We don't do this kind of split into apps/ for the other coins. It is
easier to keep related functionality in the same place.
