+++ src/rust/bitbox02-rust-c/Cargo.toml
@@ -44,7 +44,7 @@ target-firmware = ["firmware", "platform-bitbox02", "app-bitcoin", "app-litecoin
 target-firmware-btc = ["firmware", "platform-bitbox02", "app-bitcoin"]
 target-factory-setup = ["firmware", "platform-bitbox02"]
 # add Rust features which are called in the C unit tests (currently there is only one target for C tests).
-target-c-unit-tests = ["app-ethereum", "firmware"]
+target-c-unit-tests = ["app-bitcoin", "app-ethereum", "firmware"]
 
 platform-bitbox02 = []
