+++ src/rust/bitbox02-rust-c/src/app_bitcoin.rs
@@ -12,12 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use core::fmt::Write;
-
 use util::c_types::size_t;
 
-use bitbox02_rust::apps::bitcoin::keypath;
-use bitbox02_rust::apps::bitcoin::util::format_amount;
+use bitbox02_rust::hww::api::bitcoin::keypath;
 
 /// # Safety
 /// `keypath` must be not NULL and contain `keypath_len` u32 elements.
@@ -29,12 +26,7 @@ pub unsafe extern "C" fn rust_bitcoin_keypath_validate_account(
     expected_coin: u32,
 ) -> bool {
     let keypath = core::slice::from_raw_parts(keypath, keypath_len);
-    bitbox02_rust::apps::bitcoin::keypath::validate_account(
-        keypath,
-        expected_purpose,
-        expected_coin,
-    )
-    .is_ok()
+    keypath::validate_account(keypath, expected_purpose, expected_coin).is_ok()
 }
 
 /// # Safety
@@ -105,15 +97,4 @@ pub unsafe extern "C" fn rust_bitcoin_keypath_validate_address_simple(
     };
     let keypath = core::slice::from_raw_parts(keypath, keypath_len);
     keypath::validate_address_simple(keypath, expected_coin, script_type, taproot_support).is_ok()
-}
-
-/// `out` should be at least 31+len(unit) bytes.
-#[no_mangle]
-pub extern "C" fn rust_bitcoin_util_format_amount(
-    satoshi: u64,
-    unit: crate::util::CStr,
-    mut out: crate::util::CStrMut,
-) {
-    let result = format_amount(satoshi, unit.as_ref());
-    out.write_str(&result).unwrap();
 }
