+++ src/rust/bitbox02-rust-c/src/lib.rs
@@ -27,7 +27,7 @@ mod alloc;
 
 mod util;
 
-#[cfg(feature = "firmware")]
+#[cfg(any(feature = "app-bitcoin", feature = "app-litecoin"))]
 mod app_bitcoin;
 #[cfg(feature = "firmware")]
 mod async_usb;
