+++ src/rust/bitbox02-rust/src/hww.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod api;
+pub mod api;
 pub mod noise;
 
 extern crate alloc;
