+++ src/rust/bitbox02-rust/src/hww/api/bitcoin/bip143.rs
