+++ src/rust/bitbox02-rust/src/hww/api/bitcoin/keypath.rs
