+++ src/rust/bitbox02-rust/src/lib.rs
@@ -23,7 +23,6 @@ mod pb {
 
 #[macro_use]
 pub mod general;
-pub mod apps;
 pub mod async_usb;
 pub mod attestation;
 pub mod bb02_async;
